requirements.txt
# Notes:
#   Use the script tools/create_venv.sh to install these dependencies on helios

flake8
pytest
pytest-cov

fire

tensorflow_gpu==2.0.0 # Keep fixed
pandas
jsonschema
tqdm
h5py
pyyaml
matplotlib
numpy
tensorboard
jupyter
scikit-learn

spacy==2.2.2 # Version from Helios

# Note: Install as follow: "pip install tokenizers==0.6.0 transformers=2.5.1"
#   This forces the use of tokenizers 0.6.0 during installation instead of 0.5.1
#   See bug https://github.com/huggingface/transformers/issues/2831
tokenizers==0.6.0
transformers==2.5.1

sacrebleu==1.4.3

# To plot models
pydot
graphviz